refactor: references to classes and minor refactors
