+++ src/main/java/net/eriagacha/gachapool/GachaPoolService.java
@@ -1,11 +1,10 @@
-package net.eriagacha.services.gachapool;
+package net.eriagacha.gachapool;
 
-import net.eriagacha.EventHandler;
-import net.eriagacha.GachaRollEvent;
+import net.eriagacha.event.EventHandler;
+import net.eriagacha.event.GachaRollEvent;
 import net.eriagacha.models.GachaObjectModel;
 import net.eriagacha.models.GachaObjectModelItem;
 import net.eriagacha.models.GachaObjectModelStatus;
-import net.eriagacha.utils.WeightedRandomBag;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.server.network.ServerPlayerEntity;
