+++ src/main/java/net/eriagacha/gachapool/WeightedRandomBag.java
@@ -1,4 +1,4 @@
-package net.eriagacha.utils;
+package net.eriagacha.gachapool;
 
 import java.util.ArrayList;
 import java.util.List;
