+++ src/main/java/net/eriagacha/telemetry/GachaTelemetryRepository.java
@@ -1,4 +1,4 @@
-package net.eriagacha.repository;
+package net.eriagacha.telemetry;
 
 import net.eriagacha.models.GachaTelemetryModel;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
