+++ src/main/java/net/eriagacha/utils/NameSpaces.java
@@ -2,10 +2,11 @@ package net.eriagacha.utils;
 
 public class NameSpaces {
 
+  private NameSpaces() {
+  }
 
   public static final String PROJECT_NAME = "eriagacha";
 
-  private NameSpaces() {
-  }
+
 }
